feat: remove broken dev banner
Fixes #263

Also contains this squashed commit:
* fix: expect success on tap target test
