+++ lib/utils/api.dart
@@ -10,7 +10,7 @@ import "package:flutter/foundation.dart";
 import "package:freezed_annotation/freezed_annotation.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
-import "constants.dart";
+import "../dart_define.gen.dart";
 
 part "api.freezed.dart";
 part "api.g.dart";
