--- lib/utils/constants.dart
@@ -1,15 +1,0 @@
-/// This library contains the constants used throughout the app.
-library;
-
-import "package:flutter/foundation.dart";
-
-import "../features/utils/domain/device_model.dart";
-
-export "../../dart_define.gen.dart";
-
-/// The current build mode.
-const BuildMode buildMode = kReleaseMode
-    ? BuildMode.release
-    : kProfileMode
-        ? BuildMode.profile
-        : BuildMode.debug;
